Namespace E2E test steps
Ahead of adding E2E tests for our bedspace features, we pull out a new step definition file for our premises features. This is to avoid further growing a single test file with all "manage" related test steps
